top: relocate a prototype to prevent compiler warning
Until the 'locate/search' provisions were added,
top avoided the need for any function prototypes
through careful source file organization.  But
the addition of the find_string function required
a prototpe for task_show, lest a massive file
reorganization be undertaken.

This commit moves the actual protype out of top.h
and places it adjacent to the caller in order to
avoid a warning when top_nls.c is compiled.

References:
commit 270e8e7eeb66b47569940f537cdad4ab46cd36be
commit d6e6a9aa38092e55f976a398899635105dd89893

Signed-off-by: Jim Warner <james.warner@comcast.net>